requirements.txt
blinker==1.6.2
click==8.1.4
colorama==0.4.6
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.7.0
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
Werkzeug==2.3.6
zipp==3.16.0
